test(git-utils): align merge guard tests with conventions
